# src/main/resources/openapi/server.json
{"openapi":"3.1.0","info":{"title":"Opal Server","description":"OPAL is an administration layer for Open Policy Agent (OPA), detecting changes to both policy and data and pushing live updates to your agents. The opal server creates a pub/sub channel clients can subscribe to (i.e: acts as coordinator). The server also tracks a git repository (via webhook) for updates to policy (or static data) and accepts continuous data update notifications via REST api, which are then pushed to clients.","version":"0.1.0"},"paths":{"/policy":{"get":{"tags":["Bundle Server"],"summary":"Get Policy","operationId":"get_policy_policy_get","parameters":[{"description":"hash of previous bundle already downloaded, server will return a diff bundle.","required":false,"schema":{"type":"string","title":"Base Hash","description":"hash of previous bundle already downloaded, server will return a diff bundle."},"name":"base_hash","in":"query"},{"required":false,"schema":{"type":"string","title":"Base Clone Path"},"name":"base_clone_path","in":"query"},{"required":false,"schema":{"type":"string","title":"Clone Subdirectory Prefix"},"name":"clone_subdirectory_prefix","in":"query"},{"required":false,"schema":{"type":"boolean","title":"Use Fixed Path"},"name":"use_fixed_path","in":"query"},{"required":false,"schema":{"items":{"type":"string"},"type":"array","title":"Path"},"name":"path","in":"query"},{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PolicyBundle"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/policy-data":{"get":{"tags":["Data Updates"],"summary":"Default All Data","description":"A fake data source configured to be fetched by the default data\nsource config.\n\nIf the user deploying OPAL did not set DATA_CONFIG_SOURCES\nproperly, OPAL clients will be hitting this route, which will\nreturn an empty dataset (empty dict).","operationId":"default_all_data_policy_data_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/data/callback_report":{"post":{"tags":["Data Updates"],"summary":"Log Client Update Report","description":"A data update callback to be called by the OPAL client after\ncompleting an update.\n\nIf the user deploying OPAL-client did not set\nOPAL_DEFAULT_UPDATE_CALLBACKS properly, this method will be\ncalled as the default callback (will simply log the report).","operationId":"log_client_update_report_data_callback_report_post","parameters":[{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataUpdateReport"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/data/config":{"get":{"tags":["Data Updates"],"summary":"Get Data Sources Config","description":"Provides OPAL clients with their base data config, meaning from\nwhere they should fetch a *complete* picture of the policy data they\nneed.\n\nClients will use this config to pull all data when they\ninitially load and when they are reconnected to server after a\nperiod of disconnection (in which they cannot receive\nincremental updates).","operationId":"get_data_sources_config_data_config_get","parameters":[{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataSourceConfig"}}}},"307":{"description":"The data source configuration is available at another location (redirect)"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Data Updates"],"summary":"Publish Data Update Event","description":"Provides data providers (i.e: one of the backend services owned by\nwhomever deployed OPAL) with the ability to push incremental policy\ndata updates to OPAL clients.\n\nEach update contains instructions on:\n- how to fetch the data\n- where should OPAL client store the data in OPA document hierarchy\n- what clients should receive the update (through topics, only clients subscribed to provided topics will be notified)","operationId":"publish_data_update_event_data_config_post","parameters":[{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/webhook":{"post":{"tags":["Github Webhook"],"summary":"Trigger Webhook","operationId":"trigger_webhook_webhook_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/token":{"post":{"tags":["Security"],"summary":"Generate New Access Token","operationId":"generate_new_access_token_token_post","parameters":[{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessToken"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pubsub_client_info":{"get":{"tags":["Pub/Sub"],"summary":"Client Info","operationId":"client_info_pubsub_client_info_get","parameters":[{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"$ref":"#/components/schemas/ClientInfo"},"type":"object","title":"Response Client Info Pubsub Client Info Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/statistics":{"get":{"tags":["Server Statistics"],"summary":"Get Statistics","description":"Route to serve server statistics.","operationId":"get_statistics_statistics_get","parameters":[{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerStats"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stats":{"get":{"tags":["Server Statistics"],"summary":"Get Stat Counts","description":"Route to serve only server and client instanace counts.","operationId":"get_stat_counts_stats_get","parameters":[{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerStatsBrief"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/loadlimit":{"get":{"tags":["Client Load Limiting"],"summary":"Loadlimit","operationId":"loadlimit_loadlimit_get","parameters":[{"required":false,"schema":{"type":"string","title":"Authorization"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AccessToken":{"properties":{"token":{"type":"string","title":"Token"},"type":{"type":"string","title":"Type","default":"bearer"},"details":{"$ref":"#/components/schemas/TokenDetails"}},"type":"object","required":["token"],"title":"AccessToken"},"AccessTokenRequest":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"type":{"allOf":[{"$ref":"#/components/schemas/PeerType"}],"description":"The peer type we generate access token for, i.e: opal client, data provider, etc.","default":"client"},"ttl":{"type":"number","format":"time-delta","title":"Ttl","description":"Token lifetime (timedelta), can accept duration in seconds or ISO_8601 format. see: https://en.wikipedia.org/wiki/ISO_8601#Durations","default":31536000.0},"claims":{"type":"object","title":"Claims","description":"extra claims to attach to the jwt","default":{}}},"type":"object","title":"AccessTokenRequest","description":"A request to generate an access token to opal server."},"ChannelStats":{"properties":{"rpc_id":{"type":"string","title":"Rpc Id"},"client_id":{"type":"string","title":"Client Id"},"topics":{"items":{"type":"string"},"type":"array","title":"Topics"}},"type":"object","required":["rpc_id","client_id","topics"],"title":"ChannelStats"},"ClientInfo":{"properties":{"client_id":{"type":"string","title":"Client Id"},"source_host":{"type":"string","title":"Source Host"},"source_port":{"type":"integer","title":"Source Port"},"connect_time":{"type":"number","title":"Connect Time"},"subscribed_topics":{"items":{"type":"string"},"type":"array","uniqueItems":true,"title":"Subscribed Topics","default":[]},"refcount":{"type":"integer","title":"Refcount","default":0},"query_params":{"additionalProperties":{"type":"string"},"type":"object","title":"Query Params"}},"type":"object","required":["client_id","connect_time","query_params"],"title":"ClientInfo"},"DataEntryReport":{"properties":{"entry":{"allOf":[{"$ref":"#/components/schemas/DataSourceEntry"}],"title":"Entry","description":"The entry that was processed"},"fetched":{"type":"boolean","title":"Fetched","default":false},"saved":{"type":"boolean","title":"Saved","default":false},"hash":{"type":"string","title":"Hash"}},"type":"object","required":["entry"],"title":"DataEntryReport","description":"A report of the processing of a single DataSourceEntry."},"DataModule":{"properties":{"path":{"type":"string","title":"Path","description":"where to place the data module relative to opa data root"},"data":{"type":"string","title":"Data","description":"data module file contents (json)"}},"type":"object","required":["path","data"],"title":"DataModule"},"DataSourceConfig":{"properties":{"entries":{"items":{"$ref":"#/components/schemas/DataSourceEntryWithPollingInterval"},"type":"array","title":"Entries","description":"list of data sources and how to fetch from them","default":[]}},"type":"object","title":"DataSourceConfig","description":"Static list of Data Source Entries returned to client.\n\nAnswers this question for the client: from where should i get the\nfull picture of data i need? (as opposed to incremental data\nupdates)"},"DataSourceEntry":{"properties":{"url":{"type":"string","title":"Url","description":"Url source to query for data"},"config":{"type":"object","title":"Config","description":"Suggested fetcher configuration (e.g. auth or method) to fetch data with"},"topics":{"items":{"type":"string"},"type":"array","title":"Topics","description":"topics the data applies to","default":["policy_data"]},"dst_path":{"type":"string","title":"Dst Path","description":"OPA data api path to store the document at","default":""},"save_method":{"type":"string","title":"Save Method","description":"Method used to write into OPA - PUT/PATCH, when using the PATCH method the data field should conform to the JSON patch schema defined in RFC 6902(https://datatracker.ietf.org/doc/html/rfc6902#section-3)","default":"PUT"},"data":{"anyOf":[{"items":{"$ref":"#/components/schemas/JSONPatchAction"},"type":"array"},{"items":{},"type":"array"},{"type":"object"}],"title":"Data","description":"Data payload to embed within the data update (instead of having the client fetch it from the url)."}},"type":"object","required":["url"],"title":"DataSourceEntry","description":"Data source configuration - where client's should retrieve data from and how they should store it"},"DataSourceEntryWithPollingInterval":{"properties":{"url":{"type":"string","title":"Url","description":"Url source to query for data"},"config":{"type":"object","title":"Config","description":"Suggested fetcher configuration (e.g. auth or method) to fetch data with"},"topics":{"items":{"type":"string"},"type":"array","title":"Topics","description":"topics the data applies to","default":["policy_data"]},"dst_path":{"type":"string","title":"Dst Path","description":"OPA data api path to store the document at","default":""},"save_method":{"type":"string","title":"Save Method","description":"Method used to write into OPA - PUT/PATCH, when using the PATCH method the data field should conform to the JSON patch schema defined in RFC 6902(https://datatracker.ietf.org/doc/html/rfc6902#section-3)","default":"PUT"},"data":{"anyOf":[{"items":{"$ref":"#/components/schemas/JSONPatchAction"},"type":"array"},{"items":{},"type":"array"},{"type":"object"}],"title":"Data","description":"Data payload to embed within the data update (instead of having the client fetch it from the url)."},"periodic_update_interval":{"type":"number","title":"Periodic Update Interval","description":"Polling interval to refresh data from data source"}},"type":"object","required":["url"],"title":"DataSourceEntryWithPollingInterval","description":"Data source configuration - where client's should retrieve data from and how they should store it"},"DataUpdate":{"properties":{"id":{"type":"string","title":"Id"},"entries":{"items":{"$ref":"#/components/schemas/DataSourceEntry"},"type":"array","title":"Entries","description":"list of related updates the OPAL client should perform"},"reason":{"type":"string","title":"Reason","description":"Reason for triggering the update"},"callback":{"allOf":[{"$ref":"#/components/schemas/UpdateCallback"}],"title":"Callback","default":{"callbacks":[]}}},"type":"object","required":["entries"],"title":"DataUpdate","description":"DataSources used as OPAL-server configuration Data update sent to\nclients."},"DataUpdateReport":{"properties":{"update_id":{"type":"string","title":"Update Id"},"reports":{"items":{"$ref":"#/components/schemas/DataEntryReport"},"type":"array","title":"Reports"},"policy_hash":{"type":"string","title":"Policy Hash"},"user_data":{"type":"object","title":"User Data","default":{}}},"type":"object","required":["reports"],"title":"DataUpdateReport"},"DeletedFiles":{"properties":{"data_modules":{"items":{"type":"string","format":"path"},"type":"array","title":"Data Modules","default":[]},"policy_modules":{"items":{"type":"string","format":"path"},"type":"array","title":"Policy Modules","default":[]}},"type":"object","title":"DeletedFiles"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HttpFetcherConfig":{"properties":{"fetcher":{"type":"string","title":"Fetcher","description":"indicates to OPAL client that it should use a custom FetcherProvider to fetch the data"},"headers":{"type":"object","title":"Headers"},"is_json":{"type":"boolean","title":"Is Json","default":true},"process_data":{"type":"boolean","title":"Process Data","default":true},"method":{"allOf":[{"$ref":"#/components/schemas/HttpMethods"}],"default":"get"},"data":{"title":"Data"}},"type":"object","title":"HttpFetcherConfig","description":"Config for HttpFetchProvider's Adding HTTP headers."},"HttpMethods":{"enum":["get","post","put","patch","head","delete"],"title":"HttpMethods","description":"An enumeration."},"JSONPatchAction":{"properties":{"op":{"type":"string","title":"Op","description":"patch action to perform"},"path":{"type":"string","title":"Path","description":"target location in modified json"},"value":{"title":"Value","description":"json document, the operand of the action"},"from":{"type":"string","title":"From","description":"source location in json"}},"type":"object","required":["op","path"],"title":"JSONPatchAction","description":"Abstract base class for JSON patch actions (RFC 6902)"},"PeerType":{"type":"string","enum":["client","datasource","listener"],"title":"PeerType","description":"An enumeration."},"PolicyBundle":{"properties":{"manifest":{"items":{"type":"string"},"type":"array","title":"Manifest"},"hash":{"type":"string","title":"Hash","description":"commit hash (debug version)"},"old_hash":{"type":"string","title":"Old Hash","description":"old commit hash (in diff bundles)"},"data_modules":{"items":{"$ref":"#/components/schemas/DataModule"},"type":"array","title":"Data Modules"},"policy_modules":{"items":{"$ref":"#/components/schemas/RegoModule"},"type":"array","title":"Policy Modules"},"deleted_files":{"$ref":"#/components/schemas/DeletedFiles"}},"type":"object","required":["manifest","hash","data_modules","policy_modules"],"title":"PolicyBundle"},"RegoModule":{"properties":{"path":{"type":"string","title":"Path","description":"path of policy module on disk, will be used to generate policy id"},"package_name":{"type":"string","title":"Package Name","description":"opa module package name"},"rego":{"type":"string","title":"Rego","description":"rego module file contents (text)"}},"type":"object","required":["path","package_name","rego"],"title":"RegoModule"},"ServerStats":{"properties":{"uptime":{"type":"string","format":"date-time","title":"Uptime","description":"uptime for this opal server worker"},"version":{"type":"string","title":"Version","description":"opal server version"},"clients":{"additionalProperties":{"items":{"$ref":"#/components/schemas/ChannelStats"},"type":"array"},"type":"object","title":"Clients","description":"connected opal clients, each client can have multiple subscriptions"},"servers":{"items":{"type":"string"},"type":"array","uniqueItems":true,"title":"Servers","description":"list of all connected opal server replicas"}},"type":"object","required":["uptime","version","clients","servers"],"title":"ServerStats"},"ServerStatsBrief":{"properties":{"uptime":{"type":"string","format":"date-time","title":"Uptime","description":"uptime for this opal server worker"},"version":{"type":"string","title":"Version","description":"opal server version"},"client_count":{"type":"integer","title":"Client Count","description":"number of connected opal clients"},"server_count":{"type":"integer","title":"Server Count","description":"number of opal server replicas"}},"type":"object","required":["uptime","version","client_count","server_count"],"title":"ServerStatsBrief"},"TokenDetails":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"type":{"allOf":[{"$ref":"#/components/schemas/PeerType"}],"description":"The peer type we generate access token for, i.e: opal client, data provider, etc.","default":"client"},"expired":{"type":"string","format":"date-time","title":"Expired"},"claims":{"type":"object","title":"Claims"}},"type":"object","required":["id","expired","claims"],"title":"TokenDetails"},"UpdateCallback":{"properties":{"callbacks":{"items":{"anyOf":[{"type":"string"},{"items":[{"type":"string"},{"$ref":"#/components/schemas/HttpFetcherConfig"}],"type":"array","maxItems":2,"minItems":2}]},"type":"array","title":"Callbacks"}},"type":"object","required":["callbacks"],"title":"UpdateCallback","description":"Configuration of callbacks upon completion of a FetchEvent Allows\nnotifying other services on the update flow.\n\nEach callback is either a URL (str) or a tuple of a url and\nHttpFetcherConfig defining how to approach the URL"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}